requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2
pymongo==4.6.0

# Google Cloud
google-cloud-storage==2.10.0
google-auth==2.23.4
google-cloud-core==2.3.3

# Configuration and Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1

# Logging and Monitoring
structlog==23.2.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-mock==3.12.0

# Security (optional, for production)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Additional utilities
pathlib2==2.3.7.post1
typing-extensions==4.8.0

# AI and Legal Document Processing
google-generativeai
langextract==0.1.0
PyPDF2
json-repair